requirements.txt
Flask==2.3.2
flask-cors==4.0.0
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3
inference-sdk==0.9.5